move scoring.utils
